Replace custom deepEqual with fast-deep-equal npm package
- Added fast-deep-equal package to dependencies for better performance
- Created new src/utils/equality.ts wrapper that provides same API
- Updated all imports across the codebase to use new equality module
- Updated tests to use new implementation - all tests pass
- Removed custom src/utils/deepEqual.ts implementation

Benefits:
- Better performance with fast-deep-equal (fastest available package)
- Reduced maintenance burden by using well-tested library
- Same API maintained for backward compatibility
- All existing functionality preserved
