+++ package.json
@@ -71,6 +71,7 @@
     "@octokit/rest": "^22.0.0",
     "@types/js-yaml": "^4.0.9",
     "@types/moment": "^2.11.29",
+    "fast-deep-equal": "^3.1.3",
     "ical.js": "^2.2.1",
     "js-yaml": "^4.1.0",
     "moment": "^2.30.1",
