+++ src/services/BaseManager.ts
@@ -8,7 +8,7 @@ import { App, Vault, TFile } from "obsidian";
 import { TaskSyncSettings } from "../main";
 import * as yaml from "js-yaml";
 import { sanitizeFileName } from "../utils/fileNameSanitizer";
-import { settingsChanged } from "../utils/deepEqual";
+import { settingsChanged } from "../utils/equality";
 import {
   generateTasksBase as generateTasksBaseConfig,
   generateAreaBase as generateAreaBaseConfig,
